+++ dev-packages/e2e-tests/test-applications/supabase-nextjs/lib/initSupabaseAnon.ts
@@ -6,6 +6,9 @@ const NEXT_PUBLIC_SUPABASE_URL = 'http://localhost:54321';
 const NEXT_PUBLIC_SUPABASE_ANON_KEY =
   'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImV4cGxvcmV0ZXN0Iiwicm9sZSI6ImFub24iLCJpYXQiOjE2';
 
-export const supabaseClient = createClient(NEXT_PUBLIC_SUPABASE_URL, NEXT_PUBLIC_SUPABASE_ANON_KEY);
+export const getSupabaseClient = () => {
+  const supabaseClient = createClient(NEXT_PUBLIC_SUPABASE_URL, NEXT_PUBLIC_SUPABASE_ANON_KEY);
+  Sentry.addIntegration(Sentry.supabaseIntegration(supabaseClient));
 
-Sentry.addIntegration(Sentry.supabaseIntegration(supabaseClient));
+  return supabaseClient;
+};
