+++ dev-packages/e2e-tests/test-applications/supabase-nextjs/pages/_app.tsx
@@ -1,6 +1,8 @@
-import { supabaseClient } from '@/lib/initSupabaseAnon'
+import { getSupabaseClient } from '@/lib/initSupabaseAnon'
 import { SessionContextProvider } from '@supabase/auth-helpers-react'
 import type { AppProps } from 'next/app'
+
+const supabaseClient = getSupabaseClient()
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
