+++ dev-packages/e2e-tests/test-applications/supabase-nextjs/pages/api/list-users.ts
@@ -10,14 +10,7 @@ type Data = {
 const supabaseClient = getSupabaseClient();
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<Data>) {
-  // Note for test usage
-  // This only works once in tests as it will error if the user already exists
-  // So this should be called only once before all tests to create the user
-  const { data, error } = await supabaseClient.auth.admin.createUser({
-    email: 'test@sentry.test',
-    password: 'sentry.test',
-    email_confirm: true,
-  });
+  const { data, error } = await supabaseClient.auth.admin.listUsers();
 
   if (error) {
     console.warn('ERROR', error);
