+++ dev-packages/e2e-tests/test-applications/supabase-nextjs/sentry.server.config.ts
@@ -2,7 +2,7 @@
 // The config you add here will be used whenever the server handles a request.
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
-import * as Sentry from "@sentry/nextjs";
+import * as Sentry from '@sentry/nextjs';
 
 Sentry.init({
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
