+++ dev-packages/e2e-tests/test-applications/supabase-nextjs/supabase/seed.sql
@@ -1,0 +1,2 @@
+TRUNCATE auth.users CASCADE;
+TRUNCATE auth.identities CASCADE;
