--- dev-packages/e2e-tests/test-applications/supabase-nextjs/tests/client.performance.test.ts
@@ -1,70 +1,0 @@
-import { expect, test } from '@playwright/test';
-import { waitForTransaction } from '@sentry-internal/test-utils';
-
-test('Sends client-side Supabase db-operation spans and breadcrumbs to Sentry', async ({ page, baseURL }) => {
-  // Create test user
-  await fetch(`${baseURL}/api/create-test-user`);
-
-  const pageloadTransactionPromise = waitForTransaction('supabase-nextjs', transactionEvent => {
-    return transactionEvent?.contexts?.trace?.op === 'pageload' && transactionEvent?.transaction === '/';
-  });
-
-  await page.goto('/');
-
-  // Fill in login credentials
-  await page.locator('input[name=email]').fill('test@sentry.test');
-  await page.locator('input[name=password]').fill('sentry.test');
-  await page.locator('button[type=submit]').click();
-
-  // Wait for login to complete
-  await page.waitForSelector('button:has-text("Add")');
-
-  // Add a new todo entry
-  await page.locator('input[id=new-task-text]').fill('test');
-  await page.locator('button[id=add-task]').click();
-
-  const transactionEvent = await pageloadTransactionPromise;
-
-  expect(transactionEvent.spans).toContainEqual(
-    expect.objectContaining({
-      description: 'from(todos)',
-      op: 'db.select',
-      parent_span_id: expect.stringMatching(/[a-f0-9]{16}/),
-      span_id: expect.stringMatching(/[a-f0-9]{16}/),
-      start_timestamp: expect.any(Number),
-      status: 'ok',
-      timestamp: expect.any(Number),
-      trace_id: expect.stringMatching(/[a-f0-9]{32}/),
-      origin: 'auto.db.supabase',
-    }),
-  );
-
-  expect(transactionEvent.spans).toContainEqual({
-    data: expect.any(Object),
-    description: 'from(todos)',
-    op: 'db.insert',
-    parent_span_id: expect.stringMatching(/[a-f0-9]{16}/),
-    span_id: expect.stringMatching(/[a-f0-9]{16}/),
-    start_timestamp: expect.any(Number),
-    status: 'ok',
-    timestamp: expect.any(Number),
-    trace_id: expect.stringMatching(/[a-f0-9]{32}/),
-    origin: 'auto.db.supabase',
-  });
-
-  expect(transactionEvent.breadcrumbs).toContainEqual({
-    timestamp: expect.any(Number),
-    type: 'supabase',
-    category: 'db.select',
-    message: 'from(todos)',
-    data: expect.any(Object),
-  });
-
-  expect(transactionEvent.breadcrumbs).toContainEqual({
-    timestamp: expect.any(Number),
-    type: 'supabase',
-    category: 'db.insert',
-    message: 'from(todos)',
-    data: expect.any(Object),
-  });
-});
